move positions creation endpoint to supervision controller
